requirements.txt
discord.py[voice]
yt-dlp
python-dotenv
requests
yt-dlp-youtube-oauth2